+++ src/main/java/se/bjurr/prnfb/http/Invoker.java
@@ -1,4 +1,5 @@
-package se.bjurr.prnfb.listener;
+package se.bjurr.prnfb.http;
+
 
 public interface Invoker {
  void invoke(UrlInvoker urlInvoker);