+++ src/main/java/se/bjurr/prnfb/listener/PrnfbRenderer.java
@@ -7,10 +7,10 @@ import static com.google.common.collect.Lists.newArrayList;
 import static java.net.URLEncoder.encode;
 import static java.util.logging.Logger.getLogger;
 import static java.util.regex.Pattern.compile;
+import static se.bjurr.prnfb.http.UrlInvoker.urlInvoker;
+import static se.bjurr.prnfb.http.UrlInvoker.HTTP_METHOD.GET;
 import static se.bjurr.prnfb.listener.PrnfbRenderer.REPO_PROTOCOL.http;
 import static se.bjurr.prnfb.listener.PrnfbRenderer.REPO_PROTOCOL.ssh;
-import static se.bjurr.prnfb.listener.UrlInvoker.urlInvoker;
-import static se.bjurr.prnfb.listener.UrlInvoker.HTTP_METHOD.GET;
 
 import java.io.UnsupportedEncodingException;
 import java.util.Map;
@@ -18,6 +18,9 @@ import java.util.Set;
 import java.util.logging.Logger;
 import java.util.regex.Matcher;
 
+import se.bjurr.prnfb.http.ClientKeyStore;
+import se.bjurr.prnfb.http.Invoker;
+import se.bjurr.prnfb.http.UrlInvoker;
 import se.bjurr.prnfb.settings.PrnfbNotification;
 
 import com.atlassian.bitbucket.pull.PullRequest;
@@ -34,7 +37,7 @@ import com.google.common.base.Supplier;
 
 public class PrnfbRenderer {
  private static final Logger logger = getLogger(PrnfbRenderer.class.getName());
- private static Invoker invoker = new Invoker() {
+ private static Invoker mockedInvoker = new Invoker() {
   @Override
   public void invoke(UrlInvoker urlInvoker) {
    urlInvoker.invoke();
@@ -43,7 +46,19 @@ public class PrnfbRenderer {
 
  @VisibleForTesting
  public static void setInvoker(Invoker invoker) {
-  PrnfbRenderer.invoker = invoker;
+  PrnfbRenderer.mockedInvoker = invoker;
+ }
+
+ private static Invoker createInvoker() {
+  if (mockedInvoker != null) {
+   return mockedInvoker;
+  }
+  return new Invoker() {
+   @Override
+   public void invoke(UrlInvoker urlInvoker) {
+    urlInvoker.invoke();
+   }
+  };
  }
 
  public enum REPO_PROTOCOL {
@@ -56,7 +71,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getFromRef().getLatestCommit();
    }
   }), PULL_REQUEST_FROM_ID(new Resolver() {
@@ -64,7 +80,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getFromRef().getId();
    }
   }), PULL_REQUEST_FROM_BRANCH(new Resolver() {
@@ -72,7 +89,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getFromRef().getDisplayId();
    }
   }), PULL_REQUEST_FROM_REPO_ID(new Resolver() {
@@ -80,7 +98,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getFromRef().getRepository().getId() + "";
    }
   }), PULL_REQUEST_FROM_REPO_NAME(new Resolver() {
@@ -88,7 +107,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getFromRef().getRepository().getName() + "";
    }
   }), PULL_REQUEST_FROM_REPO_PROJECT_ID(new Resolver() {
@@ -96,7 +116,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getFromRef().getRepository().getProject().getId() + "";
    }
   }), PULL_REQUEST_FROM_REPO_PROJECT_KEY(new Resolver() {
@@ -104,7 +125,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getFromRef().getRepository().getProject().getKey();
    }
   }), PULL_REQUEST_FROM_REPO_SLUG(new Resolver() {
@@ -112,7 +134,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getFromRef().getRepository().getSlug() + "";
    }
   }), PULL_REQUEST_FROM_SSH_CLONE_URL(new Resolver() {
@@ -120,7 +143,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return cloneUrlFromRepository(ssh, pullRequest.getFromRef().getRepository(), repositoryService);
    }
   }), PULL_REQUEST_FROM_HTTP_CLONE_URL(new Resolver() {
@@ -128,7 +152,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return cloneUrlFromRepository(http, pullRequest.getFromRef().getRepository(), repositoryService);
    }
   }), PULL_REQUEST_ACTION(new Resolver() {
@@ -136,7 +161,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return prnfbPullRequestAction.getName();
    }
   }), PULL_REQUEST_URL(new Resolver() {
@@ -144,7 +170,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return getPullRequestUrl(propertiesService, pullRequest);
    }
   }), PULL_REQUEST_ID(new Resolver() {
@@ -152,7 +179,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getId() + "";
    }
   }), PULL_REQUEST_VERSION(new Resolver() {
@@ -160,7 +188,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getVersion() + "";
    }
   }), PULL_REQUEST_AUTHOR_ID(new Resolver() {
@@ -168,7 +197,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getAuthor().getUser().getId() + "";
    }
   }), PULL_REQUEST_AUTHOR_DISPLAY_NAME(new Resolver() {
@@ -176,7 +206,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getAuthor().getUser().getDisplayName();
    }
   }), PULL_REQUEST_AUTHOR_NAME(new Resolver() {
@@ -184,7 +215,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getAuthor().getUser().getName();
    }
   }), PULL_REQUEST_AUTHOR_EMAIL(new Resolver() {
@@ -192,7 +224,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getAuthor().getUser().getEmailAddress();
    }
   }), PULL_REQUEST_AUTHOR_SLUG(new Resolver() {
@@ -200,7 +233,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getAuthor().getUser().getSlug();
    }
   }), PULL_REQUEST_TO_HASH(new Resolver() {
@@ -208,7 +242,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getToRef().getLatestCommit();
    }
   }), PULL_REQUEST_TO_ID(new Resolver() {
@@ -216,7 +251,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getToRef().getId();
    }
   }), PULL_REQUEST_TO_BRANCH(new Resolver() {
@@ -224,7 +260,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getToRef().getDisplayId();
    }
   }), PULL_REQUEST_TO_REPO_ID(new Resolver() {
@@ -232,7 +269,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getToRef().getRepository().getId() + "";
    }
   }), PULL_REQUEST_TO_REPO_NAME(new Resolver() {
@@ -240,7 +278,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getToRef().getRepository().getName() + "";
    }
   }), PULL_REQUEST_TO_REPO_PROJECT_ID(new Resolver() {
@@ -248,7 +287,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getToRef().getRepository().getProject().getId() + "";
    }
   }), PULL_REQUEST_TO_REPO_PROJECT_KEY(new Resolver() {
@@ -256,7 +296,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getToRef().getRepository().getProject().getKey();
    }
   }), PULL_REQUEST_TO_REPO_SLUG(new Resolver() {
@@ -264,7 +305,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return pullRequest.getToRef().getRepository().getSlug() + "";
    }
   }), PULL_REQUEST_TO_SSH_CLONE_URL(new Resolver() {
@@ -272,7 +314,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return cloneUrlFromRepository(ssh, pullRequest.getToRef().getRepository(), repositoryService);
    }
   }), PULL_REQUEST_TO_HTTP_CLONE_URL(new Resolver() {
@@ -280,7 +323,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return cloneUrlFromRepository(http, pullRequest.getToRef().getRepository(), repositoryService);
    }
   }), PULL_REQUEST_COMMENT_TEXT(new Resolver() {
@@ -288,14 +332,16 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return getOrEmpty(variables, PULL_REQUEST_COMMENT_TEXT);
    }
   }), PULL_REQUEST_MERGE_COMMIT(new Resolver() {
    @Override
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfsPullRequestAction,
      ApplicationUser stashUser, RepositoryService repositoryService, ApplicationPropertiesService propertiesService,
-     PrnfbNotification prnfsNotification, Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     PrnfbNotification prnfsNotification, Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables,
+     ClientKeyStore clientKeyStore, boolean shouldAcceptAnyCertificate) {
     return getOrEmpty(variables, PULL_REQUEST_MERGE_COMMIT);
    }
   }), PULL_REQUEST_USER_DISPLAY_NAME(new Resolver() {
@@ -303,7 +349,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return applicationUser.getDisplayName();
    }
   }), PULL_REQUEST_USER_EMAIL_ADDRESS(new Resolver() {
@@ -311,7 +358,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction prnfbPullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return applicationUser.getEmailAddress();
    }
   }), PULL_REQUEST_USER_ID(new Resolver() {
@@ -319,7 +367,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction pullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return applicationUser.getId() + "";
    }
   }), PULL_REQUEST_USER_NAME(new Resolver() {
@@ -327,7 +376,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction pullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return applicationUser.getName();
    }
   }), PULL_REQUEST_USER_SLUG(new Resolver() {
@@ -335,7 +385,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction pullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return applicationUser.getSlug();
    }
   }), BUTTON_TRIGGER_TITLE(new Resolver() {
@@ -343,7 +394,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction pullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     return getOrEmpty(variables, BUTTON_TRIGGER_TITLE);
    }
   }), INJECTION_URL_VALUE(new Resolver() {
@@ -351,7 +403,8 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction pullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbRenderer.PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+     boolean shouldAcceptAnyCertificate) {
     if (!prnfbNotification.getInjectionUrl().isPresent()) {
      return "";
     }
@@ -362,8 +415,11 @@ public class PrnfbRenderer {
       .withProxyPort(prnfbNotification.getProxyPort()) //
       .withProxyUser(prnfbNotification.getProxyUser()) //
       .withProxyPassword(prnfbNotification.getProxyPassword())//
-      .appendBasicAuth(prnfbNotification);
-    PrnfbRenderer.invoker.invoke(urlInvoker);
+      .appendBasicAuth(prnfbNotification)//
+      .withClientKeyStore(clientKeyStore)//
+      .shouldAcceptAnyCertificate(shouldAcceptAnyCertificate);
+    createInvoker()//
+      .invoke(urlInvoker);
     String rawResponse = urlInvoker.getResponseString().trim();
     if (prnfbNotification.getInjectionUrlRegexp().isPresent()) {
      Matcher m = compile(prnfbNotification.getInjectionUrlRegexp().get()).matcher(rawResponse);
@@ -381,7 +437,7 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction pullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore, boolean shouldAcceptAnyCertificate) {
     return pullRequest.getTitle();
    }
   }), PULL_REQUEST_REVIEWERS_APPROVED_COUNT(new Resolver() {
@@ -389,7 +445,7 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction pullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore, boolean shouldAcceptAnyCertificate) {
     return Integer.toString(newArrayList(filter(pullRequest.getReviewers(), isApproved)).size());
    }
   }), PULL_REQUEST_PARTICIPANTS_APPROVED_COUNT(new Resolver() {
@@ -397,7 +453,7 @@ public class PrnfbRenderer {
    public String resolve(PullRequest pullRequest, PrnfbPullRequestAction pullRequestAction,
      ApplicationUser applicationUser, RepositoryService repositoryService,
      ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-     Map<PrnfbVariable, Supplier<String>> variables) {
+     Map<PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore, boolean shouldAcceptAnyCertificate) {
     return Integer.toString(newArrayList(filter(pullRequest.getParticipants(), isApproved)).size());
    }
   });
@@ -431,9 +487,9 @@ public class PrnfbRenderer {
   public String resolve(PullRequest pullRequest, PrnfbPullRequestAction pullRequestAction,
     ApplicationUser applicationUser, RepositoryService repositoryService,
     ApplicationPropertiesService propertiesService, PrnfbNotification prnfbNotification,
-    Map<PrnfbVariable, Supplier<String>> variables) {
+    Map<PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore, boolean shouldAcceptAnyCertificate) {
    return resolver.resolve(pullRequest, pullRequestAction, applicationUser, repositoryService, propertiesService,
-     prnfbNotification, variables);
+     prnfbNotification, variables, clientKeyStore, shouldAcceptAnyCertificate);
   }
 
   private static String getOrEmpty(Map<PrnfbVariable, Supplier<String>> variables, PrnfbVariable variable) {
@@ -447,7 +503,8 @@ public class PrnfbRenderer {
  public interface Resolver {
   String resolve(PullRequest pullRequest, PrnfbPullRequestAction pullRequestAction, ApplicationUser applicationUser,
     RepositoryService repositoryService, ApplicationPropertiesService propertiesService,
-    PrnfbNotification prnfbNotification, Map<PrnfbVariable, Supplier<String>> variables);
+    PrnfbNotification prnfbNotification, Map<PrnfbVariable, Supplier<String>> variables, ClientKeyStore clientKeyStore,
+    boolean shouldAcceptAnyCertificate);
  }
 
  private final RepositoryService repositoryService;
@@ -480,13 +537,13 @@ public class PrnfbRenderer {
   this.variables = variables;
  }
 
- public String render(String string, Boolean forUrl) {
+ public String render(String string, Boolean forUrl, ClientKeyStore clientKeyStore, boolean shouldAcceptAnyCertificate) {
   for (final PrnfbVariable variable : PrnfbVariable.values()) {
    final String regExpStr = "\\$\\{" + variable.name() + "\\}";
    if (string.contains(regExpStr.replaceAll("\\\\", ""))) {
     try {
      String resolved = variable.resolve(pullRequest, pullRequestAction, applicationUser, repositoryService,
-       propertiesService, prnfbNotification, variables);
+       propertiesService, prnfbNotification, variables, clientKeyStore, shouldAcceptAnyCertificate);
      string = string.replaceAll(regExpStr, forUrl ? encode(resolved, UTF_8.name()) : resolved);
     } catch (UnsupportedEncodingException e) {
      propagate(e);