+++ src/main/java/se/bjurr/prnfb/settings/PrnfbSettings.java
@@ -1,22 +1,33 @@
 package se.bjurr.prnfb.settings;
 
+import static com.google.common.base.Optional.fromNullable;
 import static com.google.common.base.Preconditions.checkNotNull;
+import static com.google.common.base.Strings.emptyToNull;
 import static com.google.common.collect.Lists.newArrayList;
 
 import java.util.List;
+
+import com.google.common.base.Optional;
 
 public class PrnfbSettings {
  private List<PrnfbNotification> notifications = newArrayList();
  private final List<PrnfbButton> buttons;
  private final boolean usersAllowed;
  private final boolean adminsAllowed;
+ private final String keyStore;
+ private final String keyStoreType;
+ private final String keyStorePassword;
+ private final boolean shouldAcceptAnyCertificate;
 
- public PrnfbSettings(List<PrnfbNotification> notifications, List<PrnfbButton> buttons, boolean usersAllowed,
-   boolean adminsAllowed) {
-  this.notifications = checkNotNull(notifications);
-  this.buttons = checkNotNull(buttons);
-  this.usersAllowed = usersAllowed;
-  this.adminsAllowed = adminsAllowed;
+ public PrnfbSettings(PrnfbSettingsBuilder builder) {
+  this.notifications = checkNotNull(builder.getNotifications());
+  this.buttons = checkNotNull(builder.getButtons());
+  this.usersAllowed = builder.isUsersAllowed();
+  this.adminsAllowed = builder.isAdminsAllowed();
+  this.keyStore = emptyToNull(builder.getKeyStore());
+  this.keyStoreType = builder.getKeyStoreType();
+  this.keyStorePassword = emptyToNull(builder.getKeyStorePassword());
+  this.shouldAcceptAnyCertificate = builder.shouldAcceptAnyCertificate();
  }
 
  public List<PrnfbNotification> getNotifications() {
@@ -33,5 +44,21 @@ public class PrnfbSettings {
 
  public boolean isAdminsAllowed() {
   return adminsAllowed;
+ }
+
+ public Optional<String> getKeyStore() {
+  return fromNullable(keyStore);
+ }
+
+ public String getKeyStoreType() {
+  return keyStoreType;
+ }
+
+ public Optional<String> getKeyStorePassword() {
+  return fromNullable(keyStorePassword);
+ }
+
+ public boolean shouldAcceptAnyCertificate() {
+  return shouldAcceptAnyCertificate;
  }
 }